+++ source/Messaging.Application/OutgoingMessages/AccountingPointCharacteristics/MarketEvaluationPoint.cs
@@ -114,79 +114,3 @@ public class MarketEvaluationPoint
 
     public ChildMarketEvaluationPoint ChildMktEvaluationPoint { get; }
 }
-
-public class ChildMarketEvaluationPoint
-{
-    public ChildMarketEvaluationPoint(string id, string description)
-    {
-        Id = id;
-        Description = description;
-    }
-
-    public string Id { get; }
-
-    public string Description { get; }
-}
-
-public class ParentMarketEvaluationPoint
-{
-    public ParentMarketEvaluationPoint(string id, string description)
-    {
-        Id = id;
-        Description = description;
-    }
-
-    public string Id { get; }
-
-    public string Description { get; }
-}
-
-public class Address
-{
-    public Address(string streetCode, string streetNmae, string streetNumber, string floorIdentification, string suiteNumber, string townCode, string townName, string townSection, string country, string postalCode)
-    {
-        StreetCode = streetCode;
-        StreetNmae = streetNmae;
-        StreetNumber = streetNumber;
-        FloorIdentification = floorIdentification;
-        SuiteNumber = suiteNumber;
-        TownCode = townCode;
-        TownName = townName;
-        TownSection = townSection;
-        Country = country;
-        PostalCode = postalCode;
-    }
-
-    public string StreetCode { get; }
-
-    public string StreetNmae { get; }
-
-    public string StreetNumber { get; }
-
-    public string FloorIdentification { get; }
-
-    public string SuiteNumber { get; }
-
-    public string TownCode { get; }
-
-    public string TownName { get; }
-
-    public string TownSection { get; }
-
-    public string Country { get; }
-
-    public string PostalCode { get; }
-}
-
-public class Series
-{
-    public Series(string product, string quantityMeasureUnit)
-    {
-        Product = product;
-        QuantityMeasureUnit = quantityMeasureUnit;
-    }
-
-    public string Product { get; }
-
-    public string QuantityMeasureUnit { get; }
-}
